refactor transfer page
